frontend: page re-organized
