+++ frontend/src/AdminBaseComponent.vue
